Add subgroup option.
In GitLab it's possible to have subgroups to organize repositories.
Protodep can not currently deal with this as it assumes the repo path
follows GitHub rules and only contains github.com/$organization/$repo.

This commit adds a subgroup property to the dependency schema that
allows the user to specify the subgroup or subgroups that are involved
in the git URL reference. When the subgroup property is not present,
the behavior remains unchanged so this should not be a breaking change.

Signed-off-by: Ryan Goodfellow <rgoodfel@isi.edu>